# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: ClipScribe API
  version: 1.0.0
  description: API v1 for submitting video intelligence jobs, tracking progress, and retrieving artifacts.
servers:
  - url: https://api.staging.clipscribe.com
    description: Staging
  - url: https://api.clipscribe.com
    description: Production
paths:
  /v1/jobs:
    post:
      summary: Submit a new job
      operationId: createJob
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/SubmitByUrl'
                - $ref: '#/components/schemas/SubmitByGcsUri'
      responses:
        '202':
          description: Accepted
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              examples:
                accepted:
                  summary: Accepted job
                  value:
                    job_id: "job_12345"
                    state: "QUEUED"
                    progress: { current_chunk: 0, total_chunks: 0 }
                    cost_to_date_usd: 0.0
                    schema_version: "1.0.0"
                    manifest_url: "https://storage.googleapis.com/bucket/jobs/job_12345/manifest.json"
                    created_at: "2025-08-08T18:30:00Z"
                    updated_at: "2025-08-08T18:30:00Z"
        '400':
          description: Invalid input
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                invalid_input:
                  value: { code: "invalid_input", message: "Provide either url or gcs_uri" }
        '401':
          description: Unauthorized
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                unauthorized:
                  value: { code: "invalid_input", message: "Missing or invalid bearer token" }
        '403':
          description: Forbidden
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                forbidden:
                  value: { code: "forbidden_origin", message: "Origin is not allowed by CORS policy" }
        '409':
          description: Duplicate (idempotent replay)
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                duplicate:
                  value: { code: "invalid_input", message: "Idempotency-Key has already been used" }
        '429':
          description: Rate limited
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
            Retry-After:
              description: Seconds to wait before retrying
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                rate_limited:
                  value: { code: "rate_limited", message: "Too many requests, please retry later", retry_after_seconds: 10 }
  /v1/jobs/{job_id}:
    get:
      summary: Get job status
      operationId: getJob
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job status
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              examples:
                analyzing:
                  summary: In progress (analyzing)
                  value:
                    job_id: "job_12345"
                    state: "ANALYZING"
                    progress: { current_chunk: 4, total_chunks: 6 }
                    cost_to_date_usd: 0.018
                    schema_version: "1.0.0"
                    manifest_url: "https://storage.googleapis.com/bucket/jobs/job_12345/manifest.json"
                    created_at: "2025-08-08T18:30:00Z"
                    updated_at: "2025-08-08T18:33:21Z"
        '401':
          description: Unauthorized
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/jobs/{job_id}/events:
    get:
      summary: Job events (SSE)
      operationId: getJobEvents
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Server-Sent Events stream
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE stream with `status`, `progress`, `cost`, `done`, `error` events.
              examples:
                sample:
                  summary: Example SSE stream
                  value: |-
                    event: status
                    data: {"state":"DOWNLOADING"}

                    event: progress
                    data: {"current_chunk":1,"total_chunks":6}

                    event: cost
                    data: {"usd":0.003}

                    event: done
                    data: {"job_id":"job_12345"}
        '401':
          description: Unauthorized
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/jobs/{job_id}/artifacts:
    get:
      summary: List job artifacts
      operationId: listArtifacts
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Artifact list
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  artifacts:
                    type: array
                    items: { $ref: '#/components/schemas/Artifact' }
              examples:
                example:
                  value:
                    job_id: "job_12345"
                    artifacts:
                      - id: "a1"
                        kind: "transcript_json"
                        size_bytes: 124567
                        url: "https://signed/url/transcript.json"
                      - id: "a2"
                        kind: "knowledge_graph_gexf"
                        size_bytes: 56012
                        url: "https://signed/url/knowledge_graph.gexf"
        '401':
          description: Unauthorized
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/uploads/presign:
    post:
      summary: Get pre-signed upload URL
      operationId: presignUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string }
                content_type: { type: string }
              required: [ filename, content_type ]
      responses:
        '200':
          description: Pre-signed URL
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_url: { type: string, format: uri }
                  gcs_uri: { type: string }
              examples:
                example:
                  value:
                    upload_url: "https://storage.googleapis.com/bucket/tmp/upload_signed"
                    gcs_uri: "gs://bucket/tmp/video_abc123.mp4"
        '400':
          description: Invalid input
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/estimate:
    get:
      summary: Get preflight cost/time estimate and admission decision
      operationId: estimate
      parameters:
        - in: query
          name: url
          schema: { type: string, format: uri }
        - in: query
          name: gcs_uri
          schema: { type: string }
      responses:
        '200':
          description: Estimate
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  estimated_cost_usd: { type: number }
                  estimated_duration_seconds: { type: number }
                  proposed_model: { type: string, enum: [ flash, pro ] }
                  admission: { type: string, enum: [ allowed, soft-blocked:rpm, blocked:budget ] }
                  reason: { type: string }
              examples:
                allowed:
                  value:
                    estimated_cost_usd: 0.035
                    estimated_duration_seconds: 120
                    proposed_model: "flash"
                    admission: "allowed"
                    reason: "Within plan caps"
                soft_blocked:
                  value:
                    estimated_cost_usd: 0.39
                    estimated_duration_seconds: 780
                    proposed_model: "flash"
                    admission: "soft-blocked:rpm"
                    reason: "Approaching RPM cap; retry soon"
                blocked:
                  value:
                    estimated_cost_usd: 1.25
                    estimated_duration_seconds: 1800
                    proposed_model: "pro"
                    admission: "blocked:budget"
                    reason: "Daily budget exceeded"
        '400':
          description: Invalid input
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SubmitByUrl:
      type: object
      properties:
        url: { type: string, format: uri }
        options:
          type: object
          additionalProperties: true
      required: [ url ]
    SubmitByGcsUri:
      type: object
      properties:
        gcs_uri: { type: string }
        options:
          type: object
          additionalProperties: true
      required: [ gcs_uri ]
    Job:
      type: object
      properties:
        job_id: { type: string }
        state: { type: string, enum: [ QUEUED, DOWNLOADING, TRANSCRIBING, UPLOADING, ANALYZING, WRITING_ARTIFACTS, COMPLETED, FAILED, CANCELED ] }
        progress:
          type: object
          properties:
            current_chunk: { type: integer }
            total_chunks: { type: integer }
        cost_to_date_usd: { type: number }
        schema_version: { type: string }
        manifest_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        error: { type: string, nullable: true }
    Error:
      type: object
      properties:
        code: { type: string, enum: [ budget_exceeded, invalid_input, rate_limited, forbidden_origin, unsupported_media, upstream_error ] }
        message: { type: string }
        retry_after_seconds: { type: integer, nullable: true }
    Artifact:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, enum: [ transcript_json, entities_json, relationships_json, knowledge_graph_gexf, report_md, manifest_json ] }
        size_bytes: { type: integer }
        url: { type: string, format: uri }
security:
  - bearerAuth: []